:zap: 최적화를 위한 React.memo, code-splitting,
