+++ src/components/index.ts
@@ -2,6 +2,6 @@ export { default as Banner } from './Card'
 export { default as Boxes } from './Header'
 export { default as Image } from './Image'
 export { default as Slick } from './Slick'
-export { default as Carousel } from './Slick/Carousel'
 export { default as Spinner } from './Spinner'
 export { default as withSuspense } from './withSuspense'
+export { default as Carousel } from './Carousel'
