+++ src/features/auth/components/index.ts
@@ -1,3 +1,2 @@
-export { default as LoginButtons } from './LoginButtons'
 export { default as LoginButton } from './LoginButton'
 export { default as LoggedinButtons } from './LoggedinButtons'
