+++ src/features/search/components/SearchForm/index.tsx
@@ -29,4 +29,4 @@ function SearchForm({
   )
 }
 
-export default SearchForm
+export default React.memo(SearchForm)
