+++ src/layouts/Layout/index.tsx
@@ -43,4 +43,4 @@ function Layout({
   )
 }
 
-export default Layout
+export default React.memo(Layout)
