+++ src/pages/Home/components/SearchForm/Title.tsx
@@ -4,4 +4,4 @@ function Title({ children }: React.PropsWithChildren) {
   return <h3>{children}</h3>
 }
 
-export default Title
+export default React.memo(Title)
