+++ src/pages/Serach/index.tsx
@@ -1,17 +1,16 @@
 import React, { useEffect } from 'react'
-import Card from './components/Card'
-import { useSearchForm } from '../../features/search/hooks/useSearchForm'
-import { useSearch } from '../../features/search/hooks/useSearchQuery'
-import SearchForm from './components/SearchForm'
-import { useQueryString } from '../../hooks/useQueryString'
-import { withSuspense } from '../../components/withSuspense'
+
+import { useSearch } from '@/features/search/hooks/useSearchQuery'
+import { useInternalRouter } from '@/hooks/useInternalRouter'
+import { useQueryString } from '@/hooks/useQueryString'
+
+import { withSuspense } from '@/components'
+import { SearchForm, Card } from './components'
 import SkeletonCards from './components/Skeleton/SkeletonCards'
-import { useInternalRouter } from '../../hooks/useInternalRouter'
 
 function Search() {
   const keyword = useQueryString('keyword')
   const { refetch, data } = useSearch(keyword, {})
-  const { value, handleSubmit, handleChange } = useSearchForm('')
   const { goBack } = useInternalRouter()
 
   useEffect(() => {
@@ -20,12 +19,7 @@ function Search() {
 
   return (
     <div className="ml-[15px] mt-3 w-[calc(100%-15px)]">
-      <SearchForm
-        onButtonClick={goBack}
-        onSubmit={handleSubmit}
-        value={value}
-        onChangeValue={handleChange}
-      />
+      <SearchForm onButtonClick={goBack} />
       <div className="grid grid-cols-2 w-[calc(100%-15px)]  gap-[5px] mt-4">
         {Array(10)
           .fill(data)
@@ -37,4 +31,4 @@ function Search() {
   )
 }
 
-export default withSuspense(Search, <SkeletonCards number={10} />)
+export default withSuspense(React.memo(Search), <SkeletonCards number={10} />)
